+++ packages/x-components/jest.config.js
@@ -1,3 +1,4 @@
 module.exports = {
-  preset: "@vue/cli-plugin-unit-jest/presets/typescript"
+  preset: '@vue/cli-plugin-unit-jest/presets/typescript',
+  testMatch: ['<rootDir>/**/*.spec.ts']
 };