+++ packages/x-components/src/utils/types.ts
@@ -1,0 +1,4 @@
+export type Dictionary<T = any> = Record<string, T>;
+export type PropsWithType<Type, PropType> = {
+  [Key in keyof Type]: Type[Key] extends PropType ? Key : never
+}[keyof Type];