+++ packages/x-components/CHANGELOG.md
@@ -1,1 +1,9 @@
 # Changelog
+
+## 1.0.0-alpha
+
+> EX-1563 Add `forEach` utility for iterating through object properties and values
+>
+> EX-1563 Add `Dictionary<T>` and `PropsWithType<Type, PropType>` util types
+>
+> EX-1563 Update jest config to allow tests in any directory